Add user registration and authentication features
Created UserService and UserRepository for handling user-related database operations. Also introduced UserEntity to model the user data. Added security configurations for user login, and views for login and registration UI. Other changes include the integration of Lombok for cleaner code, database configurations for PostgreSQL, and added necessary dependencies for these new settings in pom.xml.

This commit is necessary for introducing user authentication and registration functionality to the application. It also establishes the" user" as a new domain in the app, handling relevant data operations and user interface. This also sets up the spring security necessary for handling user login sessions.